+++ src/PackmanVsix/JSON/Completion/NameCompletionProvider.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.Composition;
-using System.Linq;
 using Microsoft.JSON.Core.Parser.TreeItems;
 using Microsoft.JSON.Editor.Completion;
 using Microsoft.VisualStudio.Language.Intellisense;
@@ -9,14 +8,14 @@ using Microsoft.VisualStudio.Utilities;
 namespace PackmanVsix
 {
     [Export(typeof(IJSONCompletionListProvider))]
-    [Name(nameof(VersionCompletionProvider))]
-    class VersionCompletionProvider : BaseCompletionProvider
+    [Name(nameof(NameCompletionProvider))]
+    class NameCompletionProvider : BaseCompletionProvider
     {
         static readonly StandardGlyphGroup _glyph = StandardGlyphGroup.GlyphGroupVariable;
 
         public override JSONCompletionContextType ContextType
         {
-            get { return JSONCompletionContextType.PropertyValue; }
+            get { return JSONCompletionContextType.PropertyName; }
         }
 
         protected override IEnumerable<JSONCompletionEntry> GetEntries(JSONCompletionContext context)
@@ -24,22 +23,17 @@ namespace PackmanVsix
             if (!VSPackage.Manager.Provider.IsInitialized)
                 yield break;
 
-            var member = context.ContextItem as JSONMember;
+            string parent = (context.ContextItem.Parent?.Parent as JSONMember)?.UnquotedNameText;
 
-            if (member == null || member.UnquotedNameText != "version")
+            if (parent != "packages")
                 yield break;
 
-            string name = (member.Parent?.Parent as JSONMember)?.UnquotedNameText;
-
-            if (string.IsNullOrEmpty(name))
-                yield break;
-
-            var versions = VSPackage.Manager.Provider.GetVersionsAsync(name).Result;
-            if (versions != null)
+            var names = VSPackage.Manager.Provider.GetPackageNamesAsync().Result;
+            if (names != null)
             {
-                foreach (var version in versions.Reverse())
+                foreach (var name in names)
                 {
-                    yield return new SimpleCompletionEntry(version, _glyph, context.Session);
+                    yield return new SimpleCompletionEntry(name, _glyph, context.Session);
                 }
             }
         }